requirements.txt
streamlit==1.37.1
pyyaml==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
tqdm==4.66.5

sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4